update api crud task by gemini

+++ src/modules/task/task.service.ts
@@ -15,13 +15,14 @@ import {
 import { Sequelize, where } from "sequelize";
 import { Op } from "sequelize";
 import { SearchTaskStatus, TaskAction, TaskStatus } from "src/enums/task.enum";
-import { buildDateFilter } from "src/utils/date";
+import { buildDateFilter, toUtcDate } from "src/utils/date";
 import { clearJsonString, getPagination } from "src/utils/common";
 import { GeminiService } from "src/shared/gemini/gemini.service";
-import { readPrompt } from "src/utils/gemini-prompt";
+import { readPrompt, suggestionPrompt } from "src/utils/gemini-prompt";
 import {
   DELETE_TASK,
   EXIST_TASK_TIME,
+  MISSING_CREATE_TASK_DATA,
   MISSING_TASK_DATA,
   NOT_FOUND_TASK,
 } from "src/common/messages/task.message";
@@ -30,12 +31,16 @@ import { UNEXPECTED_PROMPT } from "src/constants/base.constant";
 import { Cron, CronExpression, Interval, Timeout } from "@nestjs/schedule";
 import { MailService } from "../mail/mail.service";
 import { EmailTemplateNames } from "../mail/mail.constants";
+import moment from "moment-timezone";
+const userTimeZone = "Asia/Bangkok";
 
 @Injectable()
 export class TaskService {
   constructor(
     @InjectModel(Tasks)
     private readonly taskModel: typeof Tasks,
+    @InjectModel(Users)
+    private readonly userModel: typeof Users,
     private configService: ConfigService,
     private readonly geminiService: GeminiService,
     private MailService: MailService
@@ -45,7 +50,7 @@ export class TaskService {
     const { name, status, startDate, endDate } = createTaskDto;
     const toDoDay = new Date(startDate).getDay() + 1;
 
-    const task = await this.findTaskByDate(startDate, endDate);
+    const task = await this.findTaskByDate(startDate, endDate, userId);
 
     if (task) {
       throw new BadRequestException(EXIST_TASK_TIME);
@@ -56,16 +61,16 @@ export class TaskService {
       name: name.toLocaleLowerCase(),
       status,
       toDoDay,
-      startDate: new Date(startDate),
-      endDate: new Date(endDate),
+      startDate: toUtcDate(startDate),
+      endDate: toUtcDate(endDate),
     });
   }
 
-  async updateTask(id: number, createTaskDto: CreateTaskDto) {
+  async updateTask(id: number, createTaskDto: CreateTaskDto, userId: number) {
     const { name, status, startDate, endDate } = createTaskDto;
     const toDoDay = new Date(startDate).getDay() + 1;
 
-    const task = await this.findTaskByDate(startDate, endDate);
+    const task = await this.findTaskByDate(startDate, endDate, userId);
 
     if (task && task.id !== id) {
       throw new BadRequestException(EXIST_TASK_TIME);
@@ -76,19 +81,19 @@ export class TaskService {
         name: name.toLocaleLowerCase(),
         status,
         toDoDay,
-        startDate: new Date(startDate),
-        endDate: new Date(endDate),
+        startDate: toUtcDate(startDate),
+        endDate: toUtcDate(endDate),
       },
       { where: { id } }
     );
   }
 
-  async deleteTask(id: number) {
-    return await this.taskModel.destroy({ where: { id } });
+  async deleteTask(id: number, userId: number) {
+    return await this.taskModel.destroy({ where: { id, userId } });
   }
 
-  async getTaskById(id: number) {
-    return await this.taskModel.findByPk(id);
+  async getTaskById(id: number, userId: number) {
+    return await this.taskModel.findOne({ where: { id, userId } });
   }
 
   async getAllTasks(getAllTasksDto: GetAllTasksDto, userId: number) {
@@ -97,21 +102,26 @@ export class TaskService {
       where: {
         userId,
         startDate: {
-          [Op.gte]: new Date(startDate),
-          [Op.lte]: new Date(endDate),
+          [Op.gte]: toUtcDate(startDate),
+          [Op.lte]: toUtcDate(endDate),
         },
       },
-      order: [["startDate", "ASC"]],
+      attributes: [
+        "toDoDay",
+        [Sequelize.literal(`JSON_AGG("Tasks")`), "tasks"],
+      ],
       group: ["toDoDay"],
+      order: [["toDoDay", "ASC"]],
     });
   }
 
-  async findTaskByDate(startDate: string, endDate: string) {
+  async findTaskByDate(startDate: string, endDate: string, userId: number) {
     return await this.taskModel.findOne({
       where: {
+        userId,
         startDate: {
-          [Op.gte]: new Date(startDate),
-          [Op.lte]: new Date(endDate),
+          [Op.gte]: toUtcDate(startDate),
+          [Op.lte]: toUtcDate(endDate),
         },
       },
     });
@@ -160,7 +170,6 @@ export class TaskService {
       prompt = await readPrompt(prompt);
 
       const response = await this.geminiService.generateResponse(prompt);
-
       const clearResponse = clearJsonString(response);
       const jsonRes = JSON.parse(clearResponse);
 
@@ -175,106 +184,120 @@ export class TaskService {
         action,
         oldName,
         status,
-        oldEndDate,
         oldStartDate,
+        oldEndDate,
       } = jsonRes;
 
-      let parsedStartDate = startDate ? new Date(startDate) : null;
-      let parsedEndDate = endDate ? new Date(endDate) : null;
-      let parsedOldStartDate = oldStartDate ? new Date(oldStartDate) : null;
-      let parsedOldEndDate = oldEndDate ? new Date(oldEndDate) : null;
+      let parsedStartDate = startDate ? toUtcDate(startDate) : null;
+      let parsedEndDate = endDate ? toUtcDate(endDate) : null;
+      let parsedOldStartDate = oldStartDate ? toUtcDate(oldStartDate) : null;
+      let parsedOldEndDate = oldEndDate ? toUtcDate(oldEndDate) : null;
 
-      const toDoDay = (parsedStartDate || parsedOldStartDate)?.getDate();
+      const toDoDay =
+        parsedStartDate?.getDate() || parsedOldStartDate?.getDate();
 
-      const taskFromDb = await this.findTaskByDate(startDate, endDate);
-
-      if (action === TaskAction.CREATE) {
-        if (taskFromDb) {
-          throw new BadRequestException(EXIST_TASK_TIME);
-        }
-
-        return await this.taskModel.create({
-          userId,
-          name,
-          status: TaskStatus.OPEN,
-          toDoDay,
-          startDate: parsedStartDate,
-          endDate: parsedEndDate,
-        });
-      }
-
-      if (action === TaskAction.DELETE) {
-        throw new BadRequestException(DELETE_TASK);
-      }
-
-      if (action === TaskAction.FIND) {
-        if (!name || !parsedStartDate || !parsedEndDate) {
-          throw new BadRequestException(MISSING_TASK_DATA);
-        }
-
-        if (parsedStartDate.getTime() === parsedEndDate.getTime()) {
-          parsedEndDate = new Date(parsedStartDate);
-          parsedEndDate.setDate(parsedEndDate.getDate() + 1);
-        }
-
-        const task = await this.taskModel.findOne({
+      const [taskFromDb, taskList, user] = await Promise.all([
+        this.findTaskByDate(startDate, endDate, userId),
+        this.taskModel.findAll({
           where: {
             userId,
-            name: {
-              [Op.like]: `%${name}%`,
-            },
             startDate: {
-              ...(parsedStartDate && { [Op.gte]: parsedStartDate }),
-              ...(parsedEndDate && { [Op.lte]: parsedEndDate }),
+              [Op.gte]: new Date(startDate)?.setHours(0, 0, 0, 0),
+              [Op.lte]: new Date(startDate)?.setHours(23, 59, 59, 999),
             },
-            toDoDay,
           },
-        });
+        }),
+        this.userModel.findOne({ where: { id: userId } }),
+      ]);
 
-        if (!task) {
-          throw new BadRequestException(NOT_FOUND_TASK);
-        }
+      switch (action) {
+        case TaskAction.CREATE:
+          if (taskFromDb) {
+            const suggestPrompt = await suggestionPrompt(
+              name,
+              user.hobby,
+              taskList
+            );
 
-        return task;
-      }
+            const suggestTask = await this.geminiService.generateResponse(
+              suggestPrompt
+            );
 
-      if (action === TaskAction.UPDATE) {
-        if (!oldName || !parsedOldStartDate || !parsedOldEndDate) {
-          throw new BadRequestException(MISSING_TASK_DATA);
-        }
+            const { suggestName, startDate, endDate } = JSON.parse(
+              clearJsonString(suggestTask)
+            );
 
-        const task = await this.taskModel.findOne({
-          where: {
+            return {
+              name: suggestName,
+              startDate,
+              endDate,
+              status: TaskStatus.OPEN,
+            };
+          }
+
+          if (!name || !parsedStartDate || !parsedEndDate) {
+            throw new BadRequestException(MISSING_CREATE_TASK_DATA);
+          }
+
+          return await this.taskModel.create({
             userId,
-            name: oldName,
-            startDate: parsedOldStartDate,
-            endDate: parsedOldEndDate,
+            name,
+            status: TaskStatus.OPEN,
             toDoDay,
-          },
-        });
+            startDate: parsedStartDate,
+            endDate: parsedEndDate,
+          });
 
-        if (!task) {
-          throw new BadRequestException(NOT_FOUND_TASK);
-        }
+        case TaskAction.DELETE:
+          throw new BadRequestException(DELETE_TASK);
 
-        if (taskFromDb && task.id !== taskFromDb.id) {
-          throw new BadRequestException(EXIST_TASK_TIME);
-        }
+        case TaskAction.FIND:
+          if (!name || !parsedStartDate || !parsedEndDate) {
+            throw new BadRequestException(MISSING_TASK_DATA);
+          }
+          if (parsedStartDate.getTime() === parsedEndDate.getTime()) {
+            parsedEndDate.setDate(parsedEndDate.getDate() + 1);
+          }
+          const task = await this.taskModel.findOne({
+            where: {
+              userId,
+              name: { [Op.like]: `%${name}%` },
+              startDate: { [Op.gte]: parsedStartDate, [Op.lte]: parsedEndDate },
+              toDoDay,
+            },
+          });
+          if (!task) throw new BadRequestException(NOT_FOUND_TASK);
+          return task;
 
-        return await task.update({
-          name: name || task.name,
-          startDate: parsedStartDate || task.startDate,
-          endDate: parsedEndDate || task.endDate,
-          toDoDay,
-          status: status || task.status,
-        });
+        case TaskAction.UPDATE:
+          if (!oldName || !parsedOldStartDate || !parsedOldEndDate) {
+            throw new BadRequestException(MISSING_TASK_DATA);
+          }
+          const existingTask = await this.taskModel.findOne({
+            where: {
+              userId,
+              name: oldName,
+              startDate: parsedOldStartDate,
+              endDate: parsedOldEndDate,
+              toDoDay,
+            },
+          });
+          if (!existingTask) throw new BadRequestException(NOT_FOUND_TASK);
+          if (taskFromDb && existingTask.id !== taskFromDb.id)
+            throw new BadRequestException(EXIST_TASK_TIME);
+          return await existingTask.update({
+            name: name || existingTask.name,
+            startDate: parsedStartDate || existingTask.startDate,
+            endDate: parsedEndDate || existingTask.endDate,
+            toDoDay,
+            status: status || existingTask.status,
+          });
       }
     } catch (error) {
-      if (error.status) {
-        throw error;
-      }
-      console.error("Error in Prompt:", error);
-      throw new InternalServerErrorException("Failed to process task request");
+      console.error("Error in CRUDTaskByGemini:", error);
+      throw error.status
+        ? error
+        : new InternalServerErrorException("Failed to process task request");
     }
   }
